added code, but it doesn't work. to do next time:
when opening the app:

 ERROR  Failed to set list of favourites oases: [Error: [AsyncStorage] Passing null/undefined as value is not supported. If you want to remove value, Use .removeItem method instead.
Passed value: undefined
Passed key: @fav
]
(either because of the getFavourites, which passes an empty array.
Or because of the fetchFavourites effect, which tries to save an empty array.

 LOG  Retrieved an empty list of favourite oases
(expected since nothing has been saved yet)

when trying to add to favourites:

ERROR  Failed to add oase to list of favourite oases: [TypeError: 0, _setFavourites.default is not a function (it is Object)]